requirements.txt
streamlit==1.43.0
langchain==0.3.21
langchain-community==0.3.20
langchain-ollama==0.3.1
langchain-chroma==0.2.2
chromadb==0.6.3
pypdf==5.4.0
docx2txt==0.8
openpyxl==3.1.5
unstructured==0.17.2
python-dotenv==1.0.1
streamlit-agraph==3.0.4
